feat(quant): KIVI-faithful baseline -- per-channel keys (--quant-scheme kivi), hqq 8-bit control, chunked quant prefill
Week-19 A1 root cause of the W18 quant zero: in optimum-quanto's grouping, axis 0 is
per-TOKEN groups and axis -1 per-CHANNEL, so transformers' QuantizedCache default
quantized KEYS per-token -- the scheme one Qwen2.5 key-bias outlier channel defeats
(synthetic: 4-bit per-token destroys unit-variance channels, rel err 0.25 vs 0.02
per-channel). kvdlra.quant.kivi_cache builds the arm's cache: scheme 'token' =
upstream default (bit-identical, test-pinned) / 'kivi' = per-channel keys (token axis
edge-padded to the group; the '65588' divisor SKIP) + per-token values; backend
'quanto' (2/4-bit) or 'hqq' (1-8 bit; the 8-bit decode-path control). Quant arms now
honor --chunk (chunked prefill + residual flush = same fully-quantized decode start as
single-shot; the W18 ppl OOMs died in the single-shot prefill) and OOM rows print
allocated/peak GB. Aux scale+zero billed at the backend's measured dtype (2-bit/g64
asymptote 0.156x on bf16, was 0.1875x). --quant-axis-* flags removed (never produced a
row). scripts/pod/w19.sh MODEs a1diag (validate-first) + a1 (per-family grid),
w19_watchdog.sh (file-driven, restart-safe), hqq in the pod bootstrap.

+++ src/kvdlra/accounting.py
@@ -424,7 +424,7 @@ def quant_footprint(
     nbits: int,
     group: int = 64,
     residual_length: int = 128,
-    scale_words: int = 2,
+    scale_words: float = 2,
 ) -> Footprint:
     """KIVI-style 2/4-bit KV baseline (transformers ``QuantizedCache`` / quanto backend,
     arm ``quant-{nbits}bit``). The ``residual_length`` most-recent tokens stay verbatim
